Add module to display chats
